requirements.txt
Adafruit-Blinka==8.47.0
adafruit-circuitpython-busdevice==5.2.9
adafruit-circuitpython-connectionmanager==3.1.1
adafruit-circuitpython-dht==4.0.4
adafruit-circuitpython-requests==4.1.6
adafruit-circuitpython-typing==1.11.0
Adafruit-PlatformDetect==3.74.0
Adafruit-PureIO==1.1.11
beautifulsoup4==4.12.3
bidict==0.23.1
binho-host-adapter==0.1.6
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorzero==2.0
Flask==3.0.3
Flask-SocketIO==5.3.7
gpiozero==2.0.1
h11==0.14.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
psutil==6.0.0
pyftdi==0.55.4
pyserial==3.5
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.4
pyusb==1.2.1
requests==2.32.3
RPi.GPIO==0.7.1
rpi_ws281x==5.0.0
simple-websocket==1.0.0
soupsieve==2.6
sysv_ipc==1.1.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
wsproto==1.2.0